Move KinAccounts to its own file.  Adopt Sequence protocol.
